+++ lib/music_player/music_player_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:musbx/card_list.dart';
 import 'package:musbx/music_player/music_player.dart';
-import 'package:musbx/music_player/pitch_speed_card.dart';
+import 'package:musbx/music_player/pitch_speed_card/pitch_speed_card.dart';
 import 'package:musbx/music_player/position_card/button_panel.dart';
 import 'package:musbx/music_player/current_song_card/current_song_panel.dart';
 import 'package:musbx/music_player/loop_card/loop_buttons.dart';
