+++ lib/music_player/pitch_speed_card/circular_slider/circular_slider.dart
@@ -1,8 +1,8 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:musbx/music_player/circular_slider/painter.dart';
-import 'package:musbx/music_player/circular_slider/utils.dart';
+import 'package:musbx/music_player/pitch_speed_card/circular_slider/painter.dart';
+import 'package:musbx/music_player/pitch_speed_card/circular_slider/utils.dart';
 
 enum DraggingMode {
   none,
