+++ lib/music_player/pitch_speed_card/circular_slider/painter.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:musbx/music_player/circular_slider/utils.dart';
+import 'package:musbx/music_player/pitch_speed_card/circular_slider/utils.dart';
 
 class CircularSliderPainter extends CustomPainter {
   CircularSliderPainter({
